+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 * Add `LiabilityShift` class and `getLiabilityShift` to RiskData
+* Fixed a bug where `decision reasons` was returning an un-parsed string instead of array
 
 # 3.15.0
 * Add `AchReturnCode` to transaction object